requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.11
yarl==1.5.1
youtube-dl==2020.11.1.1 
discord.py[voice]==1.5.1
Pillow==7.1.2
requests==2.24.0
pymongo==3.11.2
pymongo[srv]
pymongo[srv]==3.11.2
dnspython==1.16.0
DateTime==4.3
lxml==4.6.2